modules/common/infrastructure/src/main/kotlin/weather/sprorowski/io/infrastructure/Json.kt
package weather.sprorowski.io.infrastructure

import com.fasterxml.jackson.databind.DeserializationFeature
import com.fasterxml.jackson.databind.ObjectMapper
import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
import com.fasterxml.jackson.module.kotlin.readValue

val mapper: ObjectMapper = jacksonObjectMapper()
    .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
    .registerModule(JavaTimeModule())

object Json {
    inline fun <reified T> decode(content: String): T = mapper.readValue(content)
    inline fun <reified T> encode(value: T): String = mapper.writeValueAsString(value)
}
